Move moderation logic out of @playhtml/common into partykit
Moderation is admin/server-only and should not ship in the public
@playhtml/common package. Relocate the pure-JSON helpers to
partykit/moderation.ts (imported relatively by the Worker and via a new
@moderation alias by the admin browser bundle), drop the index re-export,
and remove the changeset since the public package no longer changes.
