--- .changeset/moderation-records.md
@@ -1,7 +1,0 @@
----
-"@playhtml/common": patch
----
-
-Add moderation helpers (`extractRecords`, `removeRecordsByTargets`, `hashRecord`)
-for extracting reviewable text records from playhtml room data and removing them
-by content-hashed key. Used by the admin console's content-moderation flow.
