+++ tsconfig.json
@@ -23,7 +23,8 @@
       "playhtml": ["packages/playhtml/src"],
       "@playhtml/react": ["packages/react/src"],
       "@playhtml/common": ["packages/common/src"],
-      "@extension/*": ["extension/src/*"]
+      "@extension/*": ["extension/src/*"],
+      "@moderation": ["partykit/moderation.ts"]
     }
   },
   "include": [],
