+++ vite.config.site.mts
@@ -11,6 +11,7 @@ export default defineConfig({
       "@playhtml/react": path.join(__dirname, "packages/react/src"),
       playhtml: path.join(__dirname, "packages/playhtml/src/index.ts"),
       "@extension": path.join(__dirname, "extension/src"),
+      "@moderation": path.join(__dirname, "partykit/moderation.ts"),
     },
   },
   optimizeDeps: {
